Add infrastructure to validate test images and a first test that does so.
cdms_test_4.png was generated on a ubuntu machine.
cdmf_test_4_1.png was generated on mac lion and reflect the bug in cairo.
When cairo is fixed, remove this version entirely.

This is a manual cherry pick to restore the test_validate branch that what
was in next before improve_build replaced next.