make dashboards checkout correct vistrails and paraview nexts
The vistrails and paraview projects use the branch name to decide
what branch to check out. The dashboard scripts don't name
the devel checkout (master from uvcdat-devel.git) devel-master but
instead name it master so those two projects have been using the
wrong branches on the dashboards.

This change fixes the situation by inserting a checkout script that
checks out the branch in the same way that the wiki says developers
should interactively.